Change voicing to festival
I added another English voice since the voice included in the repository
is with accent and I think the voice I created is better for others.

I did it by adding a folder with the recordings of voice from a program
named Festival which is open source text to speech software, any sound
file which is not a recording I copied from the original folder.

I checked the license for the Festival program and I can use it for any
purpose, both commercial and non-commercial.

To use the added folder as the audio for the buds one can use the
instructions from the readme, so one should just build with the AUDIO
environment variable set to en_festival like so:

`AUDIO=en_festival ./build.sh`

Co-authored-by: Dom Rodriguez <shymega@shymega.org.uk>
Signed-off-by: Moshe Nahmias <moshegrey@gmail.com>